python_embedding
The simplest form of embedding Python is the use of the very high level interface. This interface is intended to execute a Python script without needing to interact with the application directly. This can for example be used to perform some operation on a file. The Py_SetProgramName() function should be called before Py_Initialize() to inform the interpreter about paths to Python run-time librariesEmbedding Python is similar to extending it, but not quite.